requirements.txt
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0
chromadb==0.5.5
pypdf==4.3.1
python-dotenv==1.0.1
pydantic-settings==2.5.2
streamlit==1.38.0
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers
langchain_huggingface